Explode bundle.go.
