+++ enterprise/internal/codeintel/stores/lsifstore/packages.go
@@ -1,0 +1,33 @@
+package lsifstore
+
+import (
+	"context"
+
+	"github.com/keegancsmith/sqlf"
+	"github.com/opentracing/opentracing-go/log"
+
+	"github.com/sourcegraph/sourcegraph/internal/observation"
+)
+
+// PackageInformation looks up package information data by identifier.
+func (s *Store) PackageInformation(ctx context.Context, bundleID int, path, packageInformationID string) (_ PackageInformationData, _ bool, err error) {
+	ctx, endObservation := s.operations.packageInformation.With(ctx, &err, observation.Args{LogFields: []log.Field{
+		log.Int("bundleID", bundleID),
+		log.String("path", path),
+		log.String("packageInformationID", packageInformationID),
+	}})
+	defer endObservation(1, observation.Args{})
+
+	documentData, exists, err := s.scanFirstDocumentData(s.Store.Query(ctx, sqlf.Sprintf(packageInformationQuery, bundleID, path)))
+	if err != nil || !exists {
+		return PackageInformationData{}, false, err
+	}
+
+	packageInformationData, exists := documentData.Document.PackageInformation[ID(packageInformationID)]
+	return packageInformationData, exists, nil
+}
+
+const packageInformationQuery = `
+-- source: enterprise/internal/codeintel/stores/lsifstore/bundle.go:PackageInformation
+SELECT dump_id, path, data FROM lsif_data_documents WHERE dump_id = %s AND path = %s LIMIT 1
+`
