+++ enterprise/internal/codeintel/stores/lsifstore/packages_test.go
@@ -1,0 +1,36 @@
+package lsifstore
+
+import (
+	"context"
+	"testing"
+
+	"github.com/google/go-cmp/cmp"
+
+	"github.com/sourcegraph/sourcegraph/internal/database/dbconn"
+	"github.com/sourcegraph/sourcegraph/internal/database/dbtesting"
+	"github.com/sourcegraph/sourcegraph/internal/observation"
+)
+
+func TestDatabasePackageInformation(t *testing.T) {
+	if testing.Short() {
+		t.Skip()
+	}
+	dbtesting.SetupGlobalTestDB(t)
+	populateTestStore(t)
+	store := NewStore(dbconn.Global, &observation.TestContext)
+
+	if actual, exists, err := store.PackageInformation(context.Background(), testBundleID, "protocol/protocol.go", "60"); err != nil {
+		t.Fatalf("unexpected error %s", err)
+	} else if !exists {
+		t.Errorf("no package information")
+	} else {
+		expected := PackageInformationData{
+			Name:    "github.com/sourcegraph/lsif-go",
+			Version: "v0.0.0-ad3507cbeb18",
+		}
+
+		if diff := cmp.Diff(expected, actual); diff != "" {
+			t.Errorf("unexpected package information (-want +got):\n%s", diff)
+		}
+	}
+}
